extract DecisionDraft legacy implementation behind feature flag
